Dramatically improve iterators, including adding (optional) rayon support

| style                     |           f32 |          f64 |
|---------------------------|--------------:|-------------:|
| internal (`count`/`fold`) |    2-4&times; | 2.5-5&times; |
| external (`next`/`for`)   | 1.25-2&times; | 1.4-2&times; |


Speed increases of the included benchmarks and examples, for the different types:

| style                     |           f32 |          f64 |
|---------------------------|--------------:|-------------:|
| internal benchmark (`count`/`fold`) |    2-4&times; | 2.5-5&times; |
| external benchmark (`next`/`for`)   | 1.25-2&times; | 1.4-2&times; |
| `all` example | 4.6&times; | - |
| `all_rev` example | 1.6&times; | - |

Additionally, `Iter` (the return value of `.upto`) now supports parallel iteration with `rayon` via the optional `rayon` feature. (Closes #9.)

The `all` and `all_par` benchmarks are serial and parallel (respectively) benchmarks that count all 4 billion `f32`s (exactly 4,278,190,081)  from -∞ to +∞.

| CPU                                 | `all` (s) | `all_par` (s) |    speedup |
|-------------------------------------|----------:|--------------:|-----------:|
| i7-8750H (6 cores, 12 hyperthreads) |       1.1 |          0.21 | 5.2&times; |
| i9-9880H (8 C, 16 HT)               |       1.0 |          0.16 | 6.3&times; |

It scales well. When restricting the number of threads rayon uses with `RAYON_NUM_THREADS=<count>` on the i9-9880H, the scaling is quite linear up to the core count, and there's minimal overhead:

![ieee754-rayon](https://user-images.githubusercontent.com/1203825/62832993-e1b0c080-bc7a-11e9-92c2-69718483c84d.png)

(I suspect CPU throttling is the major cause of the nonlinearity towards the end and the speed-up being not precisely the core count.)

+++ examples/all_ints.rs
@@ -8,5 +8,5 @@ fn main() {
         - /* -0.0 */ 1
         + /* infinities */ 2;
 
-    println!("count {}", (0..expected as usize).map(criterion::black_box).count());
+    println!("count {}", (0..expected as u32).map(criterion::black_box).count());
 }
